added suggestions to move toggle function and axis mode property in plot
